+++ src/main/java/com/kard/api/resources/organizations/placements/RawPlacementsClient.java
@@ -19,10 +19,12 @@ import com.kard.api.resources.commons.errors.InvalidRequest;
 import com.kard.api.resources.commons.errors.UnauthorizedError;
 import com.kard.api.resources.commons.types.ErrorResponse;
 import com.kard.api.resources.internalorganizations.types.DeleteResourceResponse;
+import com.kard.api.resources.organizations.placements.requests.GetPlacementRequest;
 import com.kard.api.resources.organizations.placements.requests.ListPlacementsRequest;
 import com.kard.api.resources.organizations.placements.types.CreatePlacementRequestBody;
 import com.kard.api.resources.organizations.placements.types.PlacementFormatUnion;
 import com.kard.api.resources.organizations.placements.types.PlacementListResponse;
+import com.kard.api.resources.organizations.placements.types.PlacementResource;
 import com.kard.api.resources.organizations.placements.types.UpdatePlacementRequestBody;
 import java.io.IOException;
 import okhttp3.Headers;
@@ -162,6 +164,10 @@ public class RawPlacementsClient {
                     request.getFilterContentStrategyId().get(),
                     false);
         }
+        if (request.getInclude().isPresent()) {
+            QueryStringMapper.addQueryParameter(
+                    httpUrl, "include", request.getInclude().get(), false);
+        }
         if (request.getPageAfter().isPresent()) {
             QueryStringMapper.addQueryParameter(
                     httpUrl, "page[after]", request.getPageAfter().get(), false);
@@ -224,32 +230,52 @@ public class RawPlacementsClient {
     /**
      * Retrieve a specific placement
      */
-    public KardApiHttpResponse<PlacementFormatUnion> get(String organizationId, String placementId) {
-        return get(organizationId, placementId, null);
+    public KardApiHttpResponse<PlacementResource> get(String organizationId, String placementId) {
+        return get(organizationId, placementId, GetPlacementRequest.builder().build());
     }
 
     /**
      * Retrieve a specific placement
      */
-    public KardApiHttpResponse<PlacementFormatUnion> get(
+    public KardApiHttpResponse<PlacementResource> get(
             String organizationId, String placementId, RequestOptions requestOptions) {
+        return get(organizationId, placementId, GetPlacementRequest.builder().build(), requestOptions);
+    }
+
+    /**
+     * Retrieve a specific placement
+     */
+    public KardApiHttpResponse<PlacementResource> get(
+            String organizationId, String placementId, GetPlacementRequest request) {
+        return get(organizationId, placementId, request, null);
+    }
+
+    /**
+     * Retrieve a specific placement
+     */
+    public KardApiHttpResponse<PlacementResource> get(
+            String organizationId, String placementId, GetPlacementRequest request, RequestOptions requestOptions) {
         HttpUrl.Builder httpUrl = HttpUrl.parse(this.clientOptions.environment().getUrl())
                 .newBuilder()
                 .addPathSegments("v2/issuers")
                 .addPathSegment(organizationId)
                 .addPathSegments("placements")
                 .addPathSegment(placementId);
+        if (request.getInclude().isPresent()) {
+            QueryStringMapper.addQueryParameter(
+                    httpUrl, "include", request.getInclude().get(), false);
+        }
         if (requestOptions != null) {
             requestOptions.getQueryParameters().forEach((_key, _value) -> {
                 httpUrl.addQueryParameter(_key, _value);
             });
         }
-        Request okhttpRequest = new Request.Builder()
+        Request.Builder _requestBuilder = new Request.Builder()
                 .url(httpUrl.build())
                 .method("GET", null)
                 .headers(Headers.of(clientOptions.headers(requestOptions)))
-                .addHeader("Accept", "application/json")
-                .build();
+                .addHeader("Accept", "application/json");
+        Request okhttpRequest = _requestBuilder.build();
         OkHttpClient client = clientOptions.httpClient();
         if (requestOptions != null && requestOptions.getTimeout().isPresent()) {
             client = clientOptions.httpClientWithTimeout(requestOptions);
@@ -259,7 +285,7 @@ public class RawPlacementsClient {
             String responseBodyString = responseBody != null ? responseBody.string() : "{}";
             if (response.isSuccessful()) {
                 return new KardApiHttpResponse<>(
-                        ObjectMappers.JSON_MAPPER.readValue(responseBodyString, PlacementFormatUnion.class), response);
+                        ObjectMappers.JSON_MAPPER.readValue(responseBodyString, PlacementResource.class), response);
             }
             try {
                 switch (response.code()) {